add params to docs, del non relevant files
