+++ meta/runtime.yml
@@ -21,7 +21,7 @@ action_groups:
     - deh_host_info
     - deh_host_type_info
     - deh_server_info
-    - dns_recordset_info
+    - dns_floating_ip_info
     - floating_ip
     - loadbalancer
     - loadbalancer_info
