+++ tests/sanity/ignore-2.10.txt
@@ -22,10 +22,10 @@ plugins/modules/cce_cluster_node.py validate-modules:missing-gplv3-license
 plugins/modules/dns_floating_ip.py validate-modules:missing-gplv3-license
 plugins/modules/dns_recordset.py validate-modules:missing-gplv3-license
 plugins/modules/dns_zone.py validate-modules:missing-gplv3-license
-plugins/modules/dns_recordset_info.py validate-modules:missing-gplv3-license
 plugins/modules/deh_host.py validate-modules:missing-gplv3-license
 plugins/modules/deh_host_info.py validate-modules:missing-gplv3-license
 plugins/modules/deh_host_type_info.py validate-modules:missing-gplv3-license
+plugins/modules/dns_floating_ip_info.py validate-modules:missing-gplv3-license
 plugins/modules/deh_server_info.py validate-modules:missing-gplv3-license
 plugins/modules/loadbalancer.py validate-modules:missing-gplv3-license
 plugins/modules/loadbalancer_info.py validate-modules:missing-gplv3-license
